refactor(source-diff-viewer): extract path-normalization, git-header-parser, binary-notice, labels
Splits source-diff-viewer.utilities.ts (898 lines) into four concern
modules per #1179: path normalization, git header parsing, binary
notice handling, and label formatting. No re-export barrel — index.ts
and source-diff-viewer.svelte now import getSourceDiffFileLabel /
getSourceDiffLineLabel from the new labels module directly.

Pure code motion; behavior and public API unchanged.
