+++ .changeset/oversized-source-diff-viewer-split.md
@@ -1,0 +1,5 @@
+---
+'@lostgradient/cinder': patch
+---
+
+Internal restructuring: split `source-diff-viewer.utilities.ts` into path-normalization, git-header-parsing, binary-notice, and label-formatting modules. No behavior or public API change.
