+++ packages/components/src/components/source-diff-viewer/index.ts
@@ -2,6 +2,7 @@ import './source-diff-viewer.css';
 import SourceDiffViewer from './source-diff-viewer.svelte';
 
 export default SourceDiffViewer;
+export { getSourceDiffFileLabel, getSourceDiffLineLabel } from './source-diff-viewer.labels.ts';
 export type {
   SourceDiffFile,
   SourceDiffHunk,
@@ -10,9 +11,5 @@ export type {
   SourceDiffParseResult,
   SourceDiffViewerProps,
 } from './source-diff-viewer.types.ts';
-export {
-  getSourceDiffFileLabel,
-  getSourceDiffLineLabel,
-  parseUnifiedPatch,
-} from './source-diff-viewer.utilities.ts';
+export { parseUnifiedPatch } from './source-diff-viewer.utilities.ts';
 export { SourceDiffViewer };
