+++ packages/components/src/components/source-diff-viewer/source-diff-viewer.utilities.ts
@@ -1,3 +1,9 @@
+import {
+  applyFileMetadata,
+  isGitBinaryNoticeMetadata,
+} from './source-diff-viewer.binary-notice.ts';
+import { createParsedGitFile } from './source-diff-viewer.git-header-parser.ts';
+import { parseGitFileSidePath } from './source-diff-viewer.path-normalization.ts';
 import type {
   SourceDiffFile,
   SourceDiffHunk,
@@ -19,194 +25,6 @@ type HunkLineResult = {
 
 const DEFAULT_MAX_LINES = 1000;
 const HUNK_HEADER_PATTERN = /^@@ -(\d+)(?:,(\d+))? \+(\d+)(?:,(\d+))? @@/;
-const TEXT_ENCODER = new TextEncoder();
-
-function stripSyntheticDiffPrefix(path: string): string {
-  return path.replace(/^[ab]\//, '');
-}
-
-function stripLeadingPathSegmentPrefix(path: string): string {
-  return path.replace(/^[^/]+\//, '');
-}
-
-function sharedPathSegmentSuffixSegments(firstPath: string, secondPath: string): string[] {
-  const firstSegments = firstPath.split('/');
-  const secondSegments = secondPath.split('/');
-  const sharedSegments: string[] = [];
-
-  while (firstSegments.length > 0 && secondSegments.length > 0) {
-    const firstSegment = firstSegments.pop();
-    const secondSegment = secondSegments.pop();
-    if (firstSegment !== secondSegment || firstSegment === undefined) break;
-    sharedSegments.unshift(firstSegment);
-  }
-
-  return sharedSegments;
-}
-
-function sharedPathSegmentSuffix(firstPath: string, secondPath: string): string | null {
-  const sharedSegments = sharedPathSegmentSuffixSegments(firstPath, secondPath);
-  return sharedSegments.length > 0 ? sharedSegments.join('/') : null;
-}
-
-function decodeGitQuotedPath(path: string): string {
-  if (!path.startsWith('"') || !path.endsWith('"')) return path;
-
-  const bytes: number[] = [];
-  const quoted = path.slice(1, -1);
-  for (let index = 0; index < quoted.length; index += 1) {
-    const character = quoted[index];
-    if (character !== '\\') {
-      const codePoint = quoted.codePointAt(index);
-      if (codePoint !== undefined) {
-        bytes.push(...TEXT_ENCODER.encode(String.fromCodePoint(codePoint)));
-        if (codePoint > 0xffff) index += 1;
-      }
-      continue;
-    }
-
-    const next = quoted[index + 1];
-    if (next && /[0-7]/.test(next)) {
-      const octal = quoted.slice(index + 1).match(/^[0-7]{1,3}/)?.[0] ?? '';
-      bytes.push(Number.parseInt(octal, 8));
-      index += octal.length;
-      continue;
-    }
-
-    const escapes: Record<string, number> = {
-      '"': 0x22,
-      '\\': 0x5c,
-      a: 0x07,
-      b: 0x08,
-      f: 0x0c,
-      n: 0x0a,
-      r: 0x0d,
-      t: 0x09,
-      v: 0x0b,
-    };
-    bytes.push(next ? (escapes[next] ?? next.charCodeAt(0)) : 0x5c);
-    if (next) index += 1;
-  }
-
-  return new TextDecoder().decode(new Uint8Array(bytes));
-}
-
-function parsePatchPath(path: string, options: { stripSyntheticPrefix: boolean }): string | null {
-  const [pathWithoutTimestamp = path] = path.split('\t');
-  const trimmedPath = decodeGitQuotedPath(pathWithoutTimestamp);
-  const normalized = options.stripSyntheticPrefix
-    ? stripSyntheticDiffPrefix(trimmedPath)
-    : trimmedPath;
-  return normalized === '/dev/null' ? null : normalized;
-}
-
-function normalizeGitHeaderPaths(
-  oldPath: string | null,
-  newPath: string | null,
-): Pick<SourceDiffFile, 'oldPath' | 'newPath'> {
-  if (!oldPath || !newPath) return { oldPath, newPath };
-  if (oldPath === newPath) return { oldPath, newPath };
-
-  const oldSegments = oldPath.split('/');
-  const newSegments = newPath.split('/');
-  const sharedSegments = sharedPathSegmentSuffixSegments(oldPath, newPath);
-  const [oldPrefix] = oldSegments;
-  const [newPrefix] = newSegments;
-  if (oldPrefix === 'a' && newPrefix === 'b') {
-    return {
-      oldPath: stripSyntheticDiffPrefix(oldPath),
-      newPath: stripSyntheticDiffPrefix(newPath),
-    };
-  }
-
-  if (oldPrefix === 'old' && newPrefix === 'new') {
-    if (
-      sharedSegments.length >= 2 &&
-      oldSegments.length === sharedSegments.length + 1 &&
-      newSegments.length === sharedSegments.length + 1
-    ) {
-      const sharedSuffix = sharedPathSegmentSuffix(oldPath, newPath);
-      return { oldPath: sharedSuffix, newPath: sharedSuffix };
-    }
-
-    return {
-      oldPath: stripLeadingPathSegmentPrefix(oldPath),
-      newPath: stripLeadingPathSegmentPrefix(newPath),
-    };
-  }
-
-  return { oldPath, newPath };
-}
-
-function readQuotedGitToken(
-  payload: string,
-  startIndex: number,
-): { token: string; endIndex: number } | null {
-  if (payload[startIndex] !== '"') return null;
-
-  for (let index = startIndex + 1; index < payload.length; index += 1) {
-    if (payload[index] === '\\') {
-      index += 1;
-      continue;
-    }
-
-    if (payload[index] === '"') {
-      return {
-        token: payload.slice(startIndex, index + 1),
-        endIndex: index + 1,
-      };
-    }
-  }
-
-  return null;
-}
-
-function parseQuotedDiffGitHeader(
-  payload: string,
-): Pick<SourceDiffFile, 'oldPath' | 'newPath'> | null {
-  const oldToken = readQuotedGitToken(payload, 0);
-  if (!oldToken) return null;
-
-  const separator = payload.slice(oldToken.endIndex).match(/^ +/);
-  if (!separator) return null;
-
-  const newToken = readQuotedGitToken(payload, oldToken.endIndex + separator[0].length);
-  if (!newToken || newToken.endIndex !== payload.length) return null;
-
-  return normalizeGitHeaderPaths(
-    parsePatchPath(oldToken.token, { stripSyntheticPrefix: false }),
-    parsePatchPath(newToken.token, { stripSyntheticPrefix: false }),
-  );
-}
-
-function parseDiffGitHeader(line: string): Pick<SourceDiffFile, 'oldPath' | 'newPath' | 'header'> {
-  const payload = line.slice('diff --git '.length);
-  const quotedHeader = parseQuotedDiffGitHeader(payload);
-  if (quotedHeader) return { ...quotedHeader, header: line };
-
-  const separatorIndexes = [...payload.matchAll(/ /g)].map((match) => match.index ?? -1);
-  for (const separatorIndex of separatorIndexes) {
-    if (separatorIndex <= 0) continue;
-    const parsedPaths = normalizeGitHeaderPaths(
-      parsePatchPath(payload.slice(0, separatorIndex), { stripSyntheticPrefix: false }),
-      parsePatchPath(payload.slice(separatorIndex + 1), { stripSyntheticPrefix: false }),
-    );
-    if (parsedPaths.oldPath && parsedPaths.oldPath === parsedPaths.newPath) {
-      return { ...parsedPaths, header: line };
-    }
-  }
-
-  const separatorIndex = separatorIndexes[separatorIndexes.length - 1] ?? -1;
-  if (separatorIndex <= 0) return { oldPath: null, newPath: null, header: line };
-
-  return {
-    ...normalizeGitHeaderPaths(
-      parsePatchPath(payload.slice(0, separatorIndex), { stripSyntheticPrefix: false }),
-      parsePatchPath(payload.slice(separatorIndex + 1), { stripSyntheticPrefix: false }),
-    ),
-    header: line,
-  };
-}
 
 function createFile(header: string | null = null): SourceDiffFile {
   return {
@@ -317,108 +135,6 @@ function createHunkMetadataLine(
   return line;
 }
 
-function parseBinaryNoticePaths(
-  file: SourceDiffFile,
-  rawLine: string,
-): Pick<SourceDiffFile, 'oldPath' | 'newPath'> | null {
-  const payload = rawLine.match(/^Binary files (.+) differ$/)?.[1];
-  if (!payload) return null;
-
-  const candidates: (Pick<SourceDiffFile, 'oldPath' | 'newPath'> & {
-    hasKnownEndpoints: boolean;
-  })[] = [];
-  let delimiterIndex = payload.indexOf(' and ');
-  while (delimiterIndex !== -1) {
-    const rawOldPath = payload.slice(0, delimiterIndex);
-    const rawNewPath = payload.slice(delimiterIndex + ' and '.length);
-    const oldPath = parsePatchPath(rawOldPath, {
-      stripSyntheticPrefix: true,
-    });
-    const newPath = parsePatchPath(rawNewPath, {
-      stripSyntheticPrefix: true,
-    });
-    const normalizedPaths = normalizeGitHeaderPaths(oldPath, newPath);
-    const normalizedOldPath =
-      normalizedPaths.oldPath !== null &&
-      file.oldPath !== null &&
-      normalizedPaths.oldPath.endsWith(`/${file.oldPath}`)
-        ? file.oldPath
-        : normalizedPaths.oldPath;
-    const normalizedNewPath =
-      normalizedPaths.newPath !== null &&
-      file.newPath !== null &&
-      normalizedPaths.newPath.endsWith(`/${file.newPath}`)
-        ? file.newPath
-        : normalizedPaths.newPath;
-    candidates.push({
-      oldPath: normalizedOldPath,
-      newPath: normalizedNewPath,
-      hasKnownEndpoints:
-        (rawOldPath.startsWith('a/') || rawOldPath === '/dev/null') &&
-        (rawNewPath.startsWith('b/') || rawNewPath === '/dev/null'),
-    });
-    delimiterIndex = payload.indexOf(' and ', delimiterIndex + 1);
-  }
-
-  const preferredCandidates = candidates.some((candidate) => candidate.hasKnownEndpoints)
-    ? candidates.filter((candidate) => candidate.hasKnownEndpoints)
-    : candidates;
-
-  if (preferredCandidates.length === 1) return preferredCandidates[0] ?? null;
-
-  return (
-    preferredCandidates.find(
-      (candidate) => candidate.oldPath === file.oldPath && candidate.newPath === file.newPath,
-    ) ?? null
-  );
-}
-
-function applyFileMetadata(file: SourceDiffFile, rawLine: string): void {
-  const binaryPaths = parseBinaryNoticePaths(file, rawLine);
-  if (binaryPaths) {
-    file.oldPath = binaryPaths.oldPath;
-    file.newPath = binaryPaths.newPath;
-    return;
-  }
-
-  if (rawLine.startsWith('new file mode ')) {
-    file.oldPath = null;
-    return;
-  }
-
-  if (rawLine.startsWith('deleted file mode ')) {
-    file.newPath = null;
-    return;
-  }
-
-  if (rawLine.startsWith('rename from ')) {
-    file.oldPath = parsePatchPath(rawLine.slice('rename from '.length), {
-      stripSyntheticPrefix: false,
-    });
-    return;
-  }
-
-  if (rawLine.startsWith('rename to ')) {
-    file.newPath = parsePatchPath(rawLine.slice('rename to '.length), {
-      stripSyntheticPrefix: false,
-    });
-    return;
-  }
-
-  if (rawLine.startsWith('copy from ')) {
-    file.oldPath = parsePatchPath(rawLine.slice('copy from '.length), {
-      stripSyntheticPrefix: false,
-    });
-    return;
-  }
-
-  if (rawLine.startsWith('copy to ')) {
-    file.newPath = parsePatchPath(rawLine.slice('copy to '.length), {
-      stripSyntheticPrefix: false,
-    });
-  }
-}
-
 function pushMetadata(
   files: SourceDiffFile[],
   rawLine: string,
@@ -505,36 +221,6 @@ function readHunkLine(
   return { renderedLineCount, lineWasRead: true, lineWasRendered: false };
 }
 
-function createParsedGitFile(rawLine: string): SourceDiffFile {
-  return {
-    ...parseDiffGitHeader(rawLine),
-    metadata: [],
-    hunks: [],
-  };
-}
-
-function parseGitFileSidePath(
-  file: SourceDiffFile,
-  rawPath: string,
-  currentPath: string | null,
-): string | null {
-  const parsedPath = parsePatchPath(rawPath, { stripSyntheticPrefix: false });
-  if (!parsedPath || !file.header?.startsWith('diff --git ')) return parsedPath;
-  if (currentPath && parsedPath === currentPath) return currentPath;
-  const [prefix] = parsedPath.split('/');
-  const hasKnownDiffPrefix =
-    prefix === 'a' || prefix === 'b' || prefix === 'old' || prefix === 'new';
-  if (prefix === 'a' || prefix === 'b') return stripSyntheticDiffPrefix(parsedPath);
-  if (currentPath && hasKnownDiffPrefix && parsedPath.endsWith(`/${currentPath}`)) {
-    return currentPath;
-  }
-
-  const pathWithoutPrefix = stripLeadingPathSegmentPrefix(parsedPath);
-  return currentPath && hasKnownDiffPrefix && pathWithoutPrefix === currentPath
-    ? currentPath
-    : parsedPath;
-}
-
 function createEmptyParseResult(): SourceDiffParseResult {
   return {
     files: [],
@@ -579,15 +265,6 @@ function isStandaloneRecursiveDiffMetadata(rawLine: string): boolean {
     (rawLine.startsWith('Symbolic links ') && rawLine.endsWith(' differ')) ||
     (rawLine.startsWith('File ') && rawLine.includes(' is a ') && rawLine.includes(' while file '))
   );
-}
-
-function isGitBinaryNoticeMetadata(file: SourceDiffFile, rawLine: string): boolean {
-  if (/^Binary files (?:a\/|\/dev\/null).+ and (?:b\/|\/dev\/null).+ differ$/.test(rawLine)) {
-    return true;
-  }
-
-  const binaryPaths = parseBinaryNoticePaths(file, rawLine);
-  return binaryPaths?.oldPath === file.oldPath && binaryPaths.newPath === file.newPath;
 }
 
 function isIndexedUnifiedFilePrelude(
@@ -644,39 +321,6 @@ function readLine(kind: SourceDiffLineKind, content: string, cursor: HunkCursor)
 function positiveInteger(value: number | undefined): number {
   if (value === undefined || !Number.isFinite(value)) return DEFAULT_MAX_LINES;
   return Math.max(0, Math.floor(value));
-}
-
-export function getSourceDiffFileLabel(file: SourceDiffFile): string {
-  if (file.newPath && file.oldPath && file.newPath !== file.oldPath) {
-    return `${file.oldPath} -> ${file.newPath}`;
-  }
-
-  return file.newPath ?? file.oldPath ?? file.header ?? 'Patch';
-}
-
-export function getSourceDiffLineLabel(line: SourceDiffLine): string {
-  if (line.kind === 'addition') {
-    return `Added line ${line.newLineNumber ?? 'unknown'}: ${line.content}`;
-  }
-
-  if (line.kind === 'removal') {
-    return `Removed line ${line.oldLineNumber ?? 'unknown'}: ${line.content}`;
-  }
-
-  if (line.kind === 'metadata') {
-    return `Diff metadata: ${line.content}`;
-  }
-
-  if (
-    line.oldLineNumber !== null &&
-    line.newLineNumber !== null &&
-    line.oldLineNumber !== line.newLineNumber
-  ) {
-    return `Context old line ${line.oldLineNumber}, new line ${line.newLineNumber}: ${line.content}`;
-  }
-
-  const lineNumber = line.newLineNumber ?? line.oldLineNumber ?? 'unknown';
-  return `Context line ${lineNumber}: ${line.content}`;
 }
 
 export function parseUnifiedPatch(
